Split into transmit & receive sockets.
We only need a single transmit socket for each address
family, but depending on the protocol we may need up to
two receive sockets.

For example, to receive TCP packets and ICMP errors, we will
need both an IPPROTO_TCP socket and an IPPROTO_ICMP socket.

This change makes this easier to implement.