Optimize dispatcher.
When sending TCP pings, we need to create a TCP raw socket. This socket will
receive all TCP packets destined for the computer, which can cause a lot of
overhead.

Optimize the dispatcher so that it should be able to discard packets it does
not care about as efficently as possible.